fix: adapt to clang21 warning options

+++ src/CHIR/Expression/Terminator.cpp
@@ -84,7 +84,7 @@ size_t Terminator::GetNumOfOperands() const
 std::vector<Value*> Terminator::GetOperands() const
 {
     CJC_ASSERT(operands.size() >= GetFirstSuccessorIndex());
-    return {operands.begin(), operands.begin() + GetFirstSuccessorIndex()};
+    return {operands.begin(), operands.begin() + static_cast<long>(GetFirstSuccessorIndex())};
 }
 
 Value* Terminator::GetOperand(size_t idx) const
@@ -288,7 +288,7 @@ std::vector<Value*> ApplyWithException::GetArgs() const
     if (GetFirstSuccessorIndex() <= 1) {
         return {};
     } else {
-        return {operands.begin() + 1, operands.begin() + GetFirstSuccessorIndex()};
+        return {operands.begin() + 1, operands.begin() + static_cast<long>(GetFirstSuccessorIndex())};
     }
 }
 
@@ -427,7 +427,7 @@ Value* InvokeWithException::GetObject() const
 /** @brief Get the call args of this InvokeWithException operation */
 std::vector<Value*> InvokeWithException::GetArgs() const
 {
-    return {operands.begin(), operands.begin() + GetFirstSuccessorIndex()};
+    return {operands.begin(), operands.begin() + static_cast<long>(GetFirstSuccessorIndex())};
 }
 
 std::string InvokeWithException::ToString([[maybe_unused]] size_t indent) const
@@ -458,7 +458,7 @@ Value* InvokeStaticWithException::GetRTTIValue() const
 
 std::vector<Value*> InvokeStaticWithException::GetArgs() const
 {
-    return {operands.begin() + 1, operands.begin() + GetFirstSuccessorIndex()};
+    return {operands.begin() + 1, operands.begin() + static_cast<long>(GetFirstSuccessorIndex())};
 }
 
 std::string InvokeStaticWithException::ToString([[maybe_unused]] size_t indent) const
